tidied, refactored and added check that the field is not mapped to a document type before deleting (fixes #2655)
git-svn-id: https://kt-dms.svn.sourceforge.net/svnroot/kt-dms/trunk@2312 c91229c3-7414-0410-bfa2-8a42b809f60b

+++ presentation/lookAndFeel/knowledgeTree/administration/docfieldmanagement/removeDocFieldBL.php
@@ -15,6 +15,7 @@ if (checkSession()) {
 	require_once("removeDocFieldUI.inc");
 	require_once("$default->fileSystemRoot/lib/security/permission.inc");
 	require_once("$default->fileSystemRoot/lib/documentmanagement/DocumentField.inc");
+	require_once("$default->fileSystemRoot/lib/documentmanagement/DocumentType.inc");
 	require_once("$default->fileSystemRoot/presentation/webpageTemplate.inc");
 	require_once("$default->fileSystemRoot/lib/visualpatterns/PatternCustom.inc");	
 	require_once("$default->fileSystemRoot/lib/foldermanagement/Folder.inc");
@@ -23,29 +24,37 @@ if (checkSession()) {
 	
 	$oPatternCustom = & new PatternCustom();	
 	
-	// get main page
 	if (isset($fDocFieldID)) {
-			
-		$oPatternCustom->setHtml(getDeletePage($fDocFieldID));
-		$main->setFormAction($_SERVER["PHP_SELF"] . "?fForDelete=1");
-		
-	// get delete page
+		$oDocField = DocumentField::get($fDocFieldID);
+		if ($oDocField) {
+					
+			// check if the document field is mapped to a document type first
+			$aDocumentTypes = $oDocField->getDocumentTypes();
+			if (count($aDocumentTypes) > 0) {
+				// display status message- can't delete
+				$oPatternCustom->setHtml(getFieldMappedPage($oDocField->getName(), $aDocumentTypes));
+			} else {
+				// perform the deletion
+				if (isset($fForDelete)) {
+					if ($oDocField->delete()) {
+						$oPatternCustom->setHtml(getDeleteSuccessPage());
+					} else {
+						$oPatternCustom->setHtml(getDeleteFailPage());
+					}
+				} else {
+					// delete confirmation page
+					$oPatternCustom->setHtml(getDeletePage($fDocFieldID));
+					$main->setFormAction($_SERVER["PHP_SELF"] . "?fForDelete=1");
+				}
+			}
+		} else {
+			// couldn't retrieve document field from db
+			$oPatternCustom->setHtml(getStatusPage("Non-existent document field", "This document field does not exist in the database"));
+		}
 	} else {
+		// prompt for a field to delete
 		$oPatternCustom->setHtml(getDeletePage(null));
 		$main->setFormAction($_SERVER["PHP_SELF"] );
-	}
-	
-		// if delete entry
-		if (isset($fForDelete)) {
-			$oDocField = DocumentField::get($fDocFieldID);
-			$oDocField->setName($fDocFieldName);
-			
-		if ($oDocField->delete()) {
-			$oPatternCustom->setHtml(getDeleteSuccessPage());
-			
-		} else {
-			$oPatternCustom->setHtml(getDeleteFailPage());
-		}
 	}
 	
 	$main->setCentralPayload($oPatternCustom);
